Refactor InputPair and Parameter classes
InputPair and Parameter classes were originally written as enumeration
classes. This requried that the index values to be hardcoded in order
for them to be convertable to integers for use with the CoolProp dll
functions. This would create a maintainability issue with requiring to
maintain the integer values of the parameters and input pairs.

Now the enumerations of the InputPair and Parameter classes have been
changed to constant properties that query the index value from the dll
functions. The values of indicies are evaluated whenever the first use
of the class is done for more information see:
https://www.mathworks.com/help/matlab/matlab_oop/expressions-in-class-definitions.html