Use default error code value in calls to coolprop
Prior to this commit, the err_code property was passed to the CoolProp
library function calls. When an error is returned and err_code is
updated to something that is not zero, a successful operation could not
change the err_code value and just return the err_code passed into the
function.

A default error code value of 0 has been added and used to all calls to
CoolProp library functions so that if a call to the library succeeds
after one that produces an error, AbstractState.err_code will reflect a
successful function call. This maintains the err_code property as an
acurate indicator if the most recent call to the CoolProp library was
successful or not.